feat: move donation management to dedicated page
