+++ src/pages/Services.tsx
@@ -120,6 +120,10 @@ const Services = () => {
     <div className="min-h-screen bg-background">
       <Header />
       <main className="container mx-auto px-4 py-8 space-y-8">
+        <div className="text-center py-6">
+          <h2 className="text-3xl font-bold text-foreground mb-2">Services à la personne</h2>
+          <p className="text-muted-foreground">Suivez vos dépenses et crédits</p>
+        </div>
         <ServicesDashboard expenses={filtered} selectedYear={selectedYear} />
 
         <div className="flex flex-col md:flex-row md:items-center md:justify-between gap-4">
